Implemented Controls Screening
There is a pause menu with a resume, controls, and quit button. The pause menu will take you to the main screen and the resume button continues the game. There is a controls viewing for people to be able to look at the controls. As well as the main menu you  can see the controls from there.